python/requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
imageio==2.9.0
numpy==1.20.3
Pillow==8.2.0
pyaml==21.10.1
PyYAML==6.0
requests==2.25.1
selenium==3.141.0
urllib3==1.26.4
